LandingHelper/requirements.txt
cffi==1.11.5
click==6.7
eventlet==0.23.0
Flask==1.0.2
Flask-SocketIO==3.0.0
Flask-SQLAlchemy==2.3.2
future==0.16.0
gevent==1.3.4
gevent-websocket==0.10.1
greenlet==0.4.13
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
pyserial==3.4
python-engineio==2.1.1
python-socketio==1.9.0
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
